final refactor to added functionality of this PR
