+++ daas/daas_app/api.py
@@ -1,6 +1,5 @@
 from rest_framework.response import Response
 from rest_framework.views import APIView
-import hashlib
 from rest_framework.parsers import MultiPartParser
 from rest_framework.parsers import JSONParser
 from rest_framework.parsers import MultiPartParser, FormParser, FileUploadParser
@@ -19,9 +18,8 @@ from rest_framework.status import (
 from .models import Sample
 from .utils.reprocess import reprocess
 from .serializers import SampleWithoutDataSerializer, SampleSerializer, ResultSerializer
-from .uploaded_files import create_and_update_file
+from .utils.new_files import create_and_upload_file
 from .utils.callback_manager import CallbackManager
-from .utils.classifier import ClassifierError
 
 
 @csrf_exempt
@@ -81,7 +79,7 @@ class UploadAPIView(APIView):
         # Upload file
         uploaded_file = request.data.get('file')
         zip_password = bytes(request.data.get('zip_password', '').encode('utf-8'))
-        file = create_and_update_file(file_name=uploaded_file.name,
+        file = create_and_upload_file(file_name=uploaded_file.name,
                                       content=uploaded_file.read(),
                                       force_reprocess=request.data.get('force_reprocess', False),
                                       zip_password=zip_password)
